requirements.txt
python-dotenv
fastapi
pydantic
google-oauth2
gspread
requests